github-action: adjust clippy check
There is no way yet to specify mutually exclusive features in Cargo, so
having `--all-features` means `mshv` and `kvm` are enabled at the same
time. That breaks clippy check.

Change the all features check to only check default features. This is a
change in behaviour.

Also add a check for `mshv`.

Signed-off-by: Wei Liu <liuwe@microsoft.com>